# front/src/api/model/accessTokenWithMeAndPagePermissions.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */
import type { Me } from "./me";
import type { PagePermissions } from "./pagePermissions";

export interface AccessTokenWithMeAndPagePermissions {
    /** the access token */
    access_token: string;
    /** the currently logged-in user info */
    me: Me;
    /** the page permissions */
    page_permissions: PagePermissions;
}