# front/src/api/model/currentPage.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */

export interface CurrentPage {
    /** the url of the page currently being viewed */
    url: string;
}