# front/src/api/model/googleLoginURL.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */

export interface GoogleLoginURL {
    /** the Google's login page URL */
    redirect_url: string;
}