# front/src/api/model/role.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */
import type { Uuid } from "./uuid";

export interface Role {
    id: Uuid;
    name: string;
    priority: number;
    created_at: string;
    updated_at: string;
}