# front/src/api/model/sortableRoleDataType.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */

export type SortableRoleDataType =
    (typeof SortableRoleDataType)[keyof typeof SortableRoleDataType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SortableRoleDataType = {
    Name: "Name",
    Priority: "Priority",
    CreatedAt: "CreatedAt",
    UpdatedAt: "UpdatedAt",
} as const;