# front/src/api/model/sortableUserDataType.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */

export type SortableUserDataType =
    (typeof SortableUserDataType)[keyof typeof SortableUserDataType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SortableUserDataType = {
    NickName: "NickName",
    LastAccess: "LastAccess",
    CreatedAt: "CreatedAt",
    UpdatedAt: "UpdatedAt",
    RoleName: "RoleName",
    RolePriority: "RolePriority",
} as const;