# front/src/api/model/user.ts
import type { Role } from "./role";
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Monitor API
 * OpenAPI spec version: 0.0.0
 */
import type { Uuid } from "./uuid";

export interface User {
    id: Uuid;
    nickname: string;
    last_access: string;
    created_at: string;
    updated_at: string;
    role: Role;
}